refactor: clean endpoint and move utils to separate file
